Removed all references to "Cotija" and replaced with "MobileControl".  Progress on multi display routing in room logic and Mobile Control bridging.
