+++ PepperDashEssentials/AppServer/Messengers/Ddvc01AtcMessenger.cs
@@ -127,7 +127,7 @@ namespace PepperDash.Essentials.AppServer.Messengers
 		/// 
 		/// </summary>
 		/// <param name="appServerController"></param>
-		protected override void CustomRegisterWithAppServer(CotijaSystemController appServerController)
+		protected override void CustomRegisterWithAppServer(MobileControlSystemController appServerController)
 		{
             //EISC.SetStringSigAction(SCurrentDialString, s => PostStatusMessage(new { currentDialString = s }));
