+++ PepperDashEssentials/Room/Types/EssentialsNDisplayRoomBase.cs
@@ -12,21 +12,21 @@ using PepperDash.Essentials.Core.Devices;
 using PepperDash.Essentials.Core.Config;
 using PepperDash.Essentials.Room.Config;
 
-namespace PepperDash.Essentials.Room.Types
+namespace PepperDash.Essentials
 {
     /// <summary>
     /// Base class for rooms with more than a single display
     /// </summary>
-    public abstract class EssentialsNDisplayRoomBase : EssentialsRoomBase
+    public abstract class EssentialsNDisplayRoomBase : EssentialsRoomBase, IHasMultipleDisplays
     {
         //public event SourceInfoChangeHandler CurrentSingleSourceChange;
 
+        public Dictionary<eSourceListItemDestinationTypes, IRoutingSinkWithSwitching> Displays { get; protected set;}
 
         public EssentialsNDisplayRoomBase(DeviceConfig config)
             : base (config)
         {
-
-            var propertiesConfig = JsonConvert.DeserializeObject<EssentialsNDisplayRoomPropertiesConfig>(config.Properties.ToString());
+            Displays = new Dictionary<eSourceListItemDestinationTypes, IRoutingSinkWithSwitching>();
 
         }
     }
