+++ essentials-framework/Essentials Core/PepperDashEssentialsBase/Routing/RoutingPort.cs
@@ -32,10 +32,10 @@ namespace PepperDash.Essentials.Core
 	public enum eRoutingSignalType
 	{
 		Audio = 1,
-		Video = 2,     
-		//AudioVideo = 4,
-        UsbOutput = 4,
-        UsbInput = 8
+		Video = 2,
+        AudioVideo = 4,
+        UsbOutput = 8,
+        UsbInput = 16
 	}
 
 	public enum eRoutingPortConnectionType
